Remove inserts by calling model methods
There had been an issue that closing the application while having multi-input
plugins in the mixer will cause a crash. The author evaded that issue with
commit f3630949.

This commit completely fixes the issue. The former commit will not be reverted
since it is also necessary even if the issue does not exist.