Use `VolumeFader` and `Meter` factories in `Mixer`
`VolumeFader` needs to be `initialize`d with sample rate and buffer size, and
`Meter` needs to be constructed with RMS window size. Many of those additional
parameters are not available in `Mixer`. Therefore, these audio devices should
be created by external codes, rather than member functions of `Mixer`.

I created this commit mainly to fix the undesired design of 1e172f01, in which
I separated the creation and initialization of `VolumeFader`, forcing me to
move the `nodeAddedCallback` to `MixerChannelListModel`, and couple `Mixer` with
`MixerChannelListModel`.